dom/devicestorage/test/test_addCorrectType.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=786922
-->
<head>
  <title>Test for basic sanity of the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=786922">Mozilla Bug 786922</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">


devicestorage_setup(function () {

function addNamed(storage, mime, fileExtension) {
  dump("adding: " + mime + " " + fileExtension + "\n");
  return navigator.getDeviceStorage(storage).addNamed(createRandomBlob(mime), randomFilename(40) + "." + fileExtension);
}

// These tests must all fail
var tests = [
  function () { return addNamed("pictures", "kyle/smash", ".png")},
  function () { return addNamed("pictures", "image/png",  ".poo")},
  function () { return addNamed("music",    "kyle/smash", ".mp3")},
  function () { return addNamed("music",    "music/mp3",  ".poo")},
  function () { return addNamed("videos",   "kyle/smash", ".ogv")},
  function () { return addNamed("videos",   "video/ogv",  ".poo")},
];

function fail(e) {
  ok(false, "addSuccess was called");
  ok(e.target.error.name == "TypeMismatchError", "Error must be TypeMismatchError");

  SimpleTest.finish();
}

function next(e) {

  if (e != undefined)
    ok(true, "addError was called");

  var f = tests.pop();

  if (f == undefined) {
    SimpleTest.finish();
    return;
  }

  request = f();
  request.onsuccess = fail;
  request.onerror = next;
}

next();

});

</script>
</pre>
</body>
</html>
